// src/shared/api/generated/models/endUserCaseEscalationResponseDtoSource.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Lola Backend Platform
 * CMS, integration, chat and realtime API for Lola AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type EndUserCaseEscalationResponseDtoSource =
  (typeof EndUserCaseEscalationResponseDtoSource)[keyof typeof EndUserCaseEscalationResponseDtoSource];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EndUserCaseEscalationResponseDtoSource = {
  END_USER_REQUEST: "END_USER_REQUEST",
  LOLA_DECISION: "LOLA_DECISION",
  CMS_USER: "CMS_USER",
  SYSTEM_POLICY: "SYSTEM_POLICY",
} as const;