// src/shared/api/generated/models/endUserCaseEscalationResponseDtoStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Retenive Backend Platform
 * CMS, integration, chat and realtime API for Retenive AI Assistant
 * OpenAPI spec version: 0.1.0
 */

export type EndUserCaseEscalationResponseDtoStatus =
  (typeof EndUserCaseEscalationResponseDtoStatus)[keyof typeof EndUserCaseEscalationResponseDtoStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const EndUserCaseEscalationResponseDtoStatus = {
  REQUESTED: "REQUESTED",
  CLAIMED: "CLAIMED",
  CLOSED: "CLOSED",
  CANCELLED: "CANCELLED",
} as const;